requirements.txt
alembic==1.15.2
Authlib==1.5.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==36.0.1
dnspython==2.7.0
email-validator==1.3.1
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.0.2
python-dotenv==0.19.0
requests==2.32.3
SQLAlchemy==2.0.19
typing_extensions==4.13.1
urllib3==2.4.0
Werkzeug==2.2.3
WTForms==3.2.1
